Updated content to conform to experiment template
Before alterations, this content was the module author-determined
description of and instructions for use of the module.
As part of an effort on the Forge team to standardize formatting and
encourage quality module documentation, templates were created for
user-experience testing. This module was chosen due to its popularity
and was changed to fit one of the template styles the Forge is testing.
Standard headings were added (Module Description, Usage,
Implementation, etc.), existent content was moved under its appropriate
heading and updated for tone, and formatting was done to adhere to
template standards.